change publish strategy
