--- src/types.d.ts
@@ -1,6 +1,0 @@
-export type Config = {
-  preload: boolean,
-  isBackground: boolean,
-  loopCount: number,
-  imagesUrls: string[],
-};
