+++ tsconfig.json
@@ -1,13 +1,19 @@
 {
   "compilerOptions": {
-    "target": "esnext",
+    "target": "es5",
     "module": "commonjs",
     "allowJs": true,
     "strict": true,
-    "moduleResolution": "node",
-    "baseUrl": "./"
+    "baseUrl": "./",
+    "declaration": true,
+    "outDir": "dist",
+    "rootDir": "src",
+    "esModuleInterop": true
   },
+  "include": ["src"],
   "exclude": [
-    "node_modules", "babel.config.js", "metro.config.js", "jest.config.js"
+    "node_modules",
+    "dist",
+    "examples"
   ]
 }
